+++ src/packs/base.js
@@ -3,17 +3,17 @@
  *
  * Essential expressions used across almost all scenarios:
  * - Data access ($get, $identity, $exists, $isPresent, $isEmpty)
- * - Basic conditionals ($if, $case, $matches, $matchesAny)
+ * - Basic conditionals ($if, $case, $matchesAll, $matchesAny)
  * - Common comparisons ($eq, $ne, $gt, $gte, $lt, $lte)
  * - Boolean logic ($and, $or, $not)
  * - Operation chaining ($pipe)
  * - Array operations ($filter, $filterBy, $map)
- * - Utilities ($default)
+ * - Utilities ($default, $debug)
  */
 
 // Import the near universal expressions
 import { $get, $identity } from "../definitions/access.js";
-import { $default, $pipe } from "../definitions/flow.js";
+import { $debug, $default, $pipe } from "../definitions/flow.js";
 import { $if, $case } from "../definitions/conditional.js";
 import {
 	$and,
@@ -25,7 +25,7 @@ import {
 	$isPresent,
 	$lt,
 	$lte,
-	$matches,
+	$matchesAll,
 	$matchesAny,
 	$ne,
 	$not,
@@ -37,6 +37,7 @@ import { $filter, $filterBy, $map } from "../definitions/array.js";
 export const base = {
 	$and,
 	$case,
+	$debug,
 	$default,
 	$eq,
 	$exists,
@@ -52,7 +53,7 @@ export const base = {
 	$lt,
 	$lte,
 	$map,
-	$matches,
+	$matchesAll,
 	$matchesAny,
 	$ne,
 	$not,